requirements.txt
torch>=2.2.0
 torchvision>=0.15.1
 opencv-python==4.7.0.72
 numpy==1.24.3
 matplotlib==3.7.1
 pillow==9.5.0